+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Migrate to loaders method for sveltekit-i18n
 - Language selection not persisted
 - Hide PokeHelp toggle when language other than English is selected
+- Show date format according to locale
 
 ## [v3.2] - 2026-03-07